# api/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exam Management API",
        "description": "API for managing exams, students, and attempt tracking",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/student/exams/{examId}": {
            "get": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Get student dashboard for an exam",
                "description": "Returns exam details, attempt status, and whether the student can start a new attempt",
                "operationId": "6605d67a6505d75bb1701af214a1ffe0",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string"
                                        },
                                        "maxAttempts": {
                                            "type": "integer"
                                        },
                                        "attemptsRemaining": {
                                            "type": "integer"
                                        },
                                        "cooldownMinutes": {
                                            "type": "integer"
                                        },
                                        "cooldownUntil": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "canStart": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/exams": {
            "get": {
                "tags": [
                    "Admin Exams"
                ],
                "summary": "List all exams",
                "operationId": "1a6bc0ecfd8688d4d884f9374f45dece",
                "responses": {
                    "200": {
                        "description": "List of exams",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "maxAttempts": {
                                                "type": "integer"
                                            },
                                            "cooldownMinutes": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Exams"
                ],
                "summary": "Create a new exam",
                "operationId": "4fd2f7f107e96cde8fa884367ba8a1a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "maxAttempts",
                                    "cooldownMinutes"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mathematics Final Exam"
                                    },
                                    "maxAttempts": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 3
                                    },
                                    "cooldownMinutes": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exam created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "examId": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/exams/{examId}": {
            "get": {
                "tags": [
                    "Admin Exams"
                ],
                "summary": "Get exam details",
                "operationId": "8d1b78e726e4a62f4aec0bc3af0c5df3",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "maxAttempts": {
                                            "type": "integer"
                                        },
                                        "cooldownMinutes": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Exams"
                ],
                "summary": "Update an exam (resets all attempts)",
                "operationId": "7157ff5d8ce82e73ab9feeed5b691f54",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "maxAttempts",
                                    "cooldownMinutes"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Exam Title"
                                    },
                                    "maxAttempts": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "cooldownMinutes": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Exam updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/exams/{examId}/attempts": {
            "get": {
                "tags": [
                    "Admin Exams"
                ],
                "summary": "Get attempt history for an exam",
                "operationId": "58e82c6753a02cddda0fbe3cb23dbc9e",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of attempts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "studentId": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "studentName": {
                                                "type": "string"
                                            },
                                            "studentEmail": {
                                                "type": "string"
                                            },
                                            "attemptNumber": {
                                                "type": "integer"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "IN_PROGRESS",
                                                    "COMPLETED"
                                                ]
                                            },
                                            "startedAt": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "endedAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}